Move DialogHost to sep file / handle all input events
